Updated golden image
See https://android-review.googlesource.com/c/platform/frameworks/support/+/1563735

I'm not really sure why this change is showing up now, I actually expected it earlier when we changed
the highlight border to 3dp.  Debugging these tests is really painful (can only be done via the CQ) so I don't feel getting to the bottom of this is worth the effort.

Change-Id: I0a285e2162361afa1903e1ed76698756067478a3