Updated goldens to go with asop/1900424
The color changes are due to setting the alpha and in the test watch face some of the colors set the alpha by mistake. See. https://cs.android.com/androidx/platform/frameworks/support/+/androidx-main:wear/watchface/watchface/samples/src/main/res/values/colors.xml?q=colors.xml%20f:watch&ss=androidx%2Fplatform%2Fframeworks%2Fsupport

The time change for directBoot is a surprise. I looked into this and the test is configured to set time to 100000 milliseconds past the epoch in UTC time. I.e. 1 minute and 40s past midnight, which is the time in the updated image.

I have a feeling we were not precise in the past re timezone in the tests and maybe this test has been failing for a long time and been ignored...

Change-Id: I78074af46a654eb4870747b11907f3a150eca871