Update golden images (auto-generated from Scuba): invocation I45200010318509277
Note: The CLs in this topic include a regression seen in these screenshots where input points are interpreted incorrectly. Fixes for this are in progress, but in order to unblock API review, the aim is to land this topic as is, send out a CL for API+2 review, and then create a follow-up topic/CL with the fixes and updated screenshot goldens.

Change-Id: I1d9e0d7ed0ffe086b6685ffc4ba5e3780bfbb07b
Flag: TEST_ONLY
Test: atest MyRoboTests